% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nhanes.R
\name{nhanes}
\alias{nhanes}
\title{Download an XPT table and return as a data frame}
\usage{
nhanes(year, nh_table)
}
\arguments{
\item{year}{The year in yyyy format where 1999 <= yyyy <= 2012}

\item{nh_table}{The name of the specific table to retrieve}
}
\value{
The table is returned as a data frame
}
\description{
Download an XPT table and return as a data frame
}
\details{
Currently, a table is downloaded in its entirety. Other than nhanesAttr, this is the
only function that actually accesses data from the nhanes site directly.
}
\examples{
nhanes(2007, 'BPX_E')
nhanes(2009, 'FOLATE_F')
}

